+++ lib/punchblock/translator/asterisk/call.rb
@@ -64,6 +64,7 @@ module Punchblock
         def to_s
           "#<#{self.class}:#{id} Channel: #{channel.inspect}>"
         end
+        alias :inspect :to_s
 
         def dial(dial_command)
           @direction = :outbound
@@ -226,8 +227,6 @@ module Punchblock
           "#{self.class}: #{id}"
         end
 
-        private
-
         def redirect_back(other_call = nil)
           redirect_options = {
             'Channel'   => channel,
@@ -243,6 +242,8 @@ module Punchblock
           }) if other_call
           send_ami_action 'Redirect', redirect_options
         end
+
+        private
 
         def send_end_event(reason)
           send_pb_event Event::End.new(:reason => reason)