+++ spec/punchblock/translator/asterisk/component/output_spec.rb
@@ -9,14 +9,14 @@ module Punchblock
         describe Output do
           let(:connection) do
             mock_connection_with_event_handler do |event|
-              command.add_event event
+              original_command.add_event event
             end
           end
           let(:media_engine)  { nil }
           let(:translator)    { Punchblock::Translator::Asterisk.new mock('AMI'), connection, media_engine }
           let(:mock_call)     { Punchblock::Translator::Asterisk::Call.new 'foo', translator }
 
-          let :command do
+          let :original_command do
             Punchblock::Component::Output.new command_options
           end
 
@@ -30,10 +30,10 @@ module Punchblock
             { :ssml => ssml_doc }
           end
 
-          subject { Output.new command, mock_call }
+          subject { Output.new original_command, mock_call }
 
           describe '#execute' do
-            before { command.request! }
+            before { original_command.request! }
 
             it "calls answer_if_not_answered on the call" do
               mock_call.expects :answer_if_not_answered
@@ -75,7 +75,7 @@ module Punchblock
                   block.call Punchblock::Component::Asterisk::AGI::Command::Complete::Success.new(:code => 200, :result => 1)
                 end
                 subject.execute
-                command.complete_event(0.1).reason.should be_a Punchblock::Component::Output::Complete::Success
+                original_command.complete_event(0.1).reason.should be_a Punchblock::Component::Output::Complete::Success
               end
 
               describe 'interrupt_on' do
@@ -108,7 +108,7 @@ module Punchblock
                   it "should return an error and not execute any actions" do
                     subject.execute
                     error = ProtocolError.new.setup 'option error', 'An interrupt-on value of speech is unsupported.'
-                    command.response(0.1).should be == error
+                    original_command.response(0.1).should be == error
                   end
                 end
               end
@@ -168,7 +168,7 @@ module Punchblock
                   block.call Punchblock::Component::Asterisk::AGI::Command::Complete::Success.new(:code => 200, :result => 1)
                 end
                 subject.execute
-                command.complete_event(0.1).reason.should be_a Punchblock::Component::Output::Complete::Success
+                original_command.complete_event(0.1).reason.should be_a Punchblock::Component::Output::Complete::Success
               end
 
               describe 'ssml' do
@@ -177,7 +177,7 @@ module Punchblock
                   it "should return an error and not execute any actions" do
                     subject.execute
                     error = ProtocolError.new.setup 'option error', 'An SSML document is required.'
-                    command.response(0.1).should be == error
+                    original_command.response(0.1).should be == error
                   end
                 end
               end
@@ -196,7 +196,7 @@ module Punchblock
                   it "should return an error and not execute any actions" do
                     subject.execute
                     error = ProtocolError.new.setup 'option error', 'A start_offset value is unsupported on Asterisk.'
-                    command.response(0.1).should be == error
+                    original_command.response(0.1).should be == error
                   end
                 end
               end
@@ -215,7 +215,7 @@ module Punchblock
                   it "should return an error and not execute any actions" do
                     subject.execute
                     error = ProtocolError.new.setup 'option error', 'A start_paused value is unsupported on Asterisk.'
-                    command.response(0.1).should be == error
+                    original_command.response(0.1).should be == error
                   end
                 end
               end
@@ -234,7 +234,7 @@ module Punchblock
                   it "should return an error and not execute any actions" do
                     subject.execute
                     error = ProtocolError.new.setup 'option error', 'A repeat_interval value is unsupported on Asterisk.'
-                    command.response(0.1).should be == error
+                    original_command.response(0.1).should be == error
                   end
                 end
               end
@@ -253,7 +253,7 @@ module Punchblock
                   it "should return an error and not execute any actions" do
                     subject.execute
                     error = ProtocolError.new.setup 'option error', 'A repeat_times value is unsupported on Asterisk.'
-                    command.response(0.1).should be == error
+                    original_command.response(0.1).should be == error
                   end
                 end
               end
@@ -272,7 +272,7 @@ module Punchblock
                   it "should return an error and not execute any actions" do
                     subject.execute
                     error = ProtocolError.new.setup 'option error', 'A max_time value is unsupported on Asterisk.'
-                    command.response(0.1).should be == error
+                    original_command.response(0.1).should be == error
                   end
                 end
               end
@@ -325,7 +325,7 @@ module Punchblock
                   it "should return an error and not execute any actions" do
                     subject.execute
                     error = ProtocolError.new.setup 'option error', 'An interrupt-on value of speech is unsupported.'
-                    command.response(0.1).should be == error
+                    original_command.response(0.1).should be == error
                   end
                 end
               end
@@ -356,7 +356,7 @@ module Punchblock
                 { :ssml => ssml_doc }.merge(command_opts)
               end
 
-              let :command do
+              let :original_command do
                 Punchblock::Component::Output.new command_options
               end
 
@@ -366,7 +366,7 @@ module Punchblock
                   it "should return an error and not execute any actions" do
                     subject.execute
                     error = ProtocolError.new.setup 'option error', 'An SSML document is required.'
-                    command.response(0.1).should be == error
+                    original_command.response(0.1).should be == error
                   end
                 end
 
@@ -388,7 +388,7 @@ module Punchblock
                       block.call Punchblock::Component::Asterisk::AGI::Command::Complete::Success.new(:code => 200, :result => 1)
                     end
                     subject.execute
-                    command.complete_event(0.1).reason.should be_a Punchblock::Component::Output::Complete::Success
+                    original_command.complete_event(0.1).reason.should be_a Punchblock::Component::Output::Complete::Success
                   end
                 end
 
@@ -410,7 +410,7 @@ module Punchblock
                       block.call Punchblock::Component::Asterisk::AGI::Command::Complete::Success.new(:code => 200, :result => 1)
                     end
                     subject.execute
-                    command.complete_event(0.1).reason.should be_a Punchblock::Component::Output::Complete::Success
+                    original_command.complete_event(0.1).reason.should be_a Punchblock::Component::Output::Complete::Success
                   end
                 end
 
@@ -445,7 +445,7 @@ module Punchblock
                     def mock_call.send_agi_action!(*args, &block)
                       block.call Punchblock::Component::Asterisk::AGI::Command::Complete::Success.new(:code => 200, :result => 1)
                     end
-                    command.expects(:add_event).once.with do |e|
+                    original_command.expects(:add_event).once.with do |e|
                       e.reason.should be_a Punchblock::Component::Output::Complete::Success
                     end
                     subject.execute
@@ -464,7 +464,7 @@ module Punchblock
                   it "should return an unrenderable document error" do
                     subject.execute
                     error = ProtocolError.new.setup 'unrenderable document error', 'The provided document could not be rendered.'
-                    command.response(0.1).should be == error
+                    original_command.response(0.1).should be == error
                   end
                 end
               end
@@ -483,7 +483,7 @@ module Punchblock
                   it "should return an error and not execute any actions" do
                     subject.execute
                     error = ProtocolError.new.setup 'option error', 'A start_offset value is unsupported on Asterisk.'
-                    command.response(0.1).should be == error
+                    original_command.response(0.1).should be == error
                   end
                 end
               end
@@ -502,7 +502,7 @@ module Punchblock
                   it "should return an error and not execute any actions" do
                     subject.execute
                     error = ProtocolError.new.setup 'option error', 'A start_paused value is unsupported on Asterisk.'
-                    command.response(0.1).should be == error
+                    original_command.response(0.1).should be == error
                   end
                 end
               end
@@ -521,7 +521,7 @@ module Punchblock
                   it "should return an error and not execute any actions" do
                     subject.execute
                     error = ProtocolError.new.setup 'option error', 'A repeat_interval value is unsupported on Asterisk.'
-                    command.response(0.1).should be == error
+                    original_command.response(0.1).should be == error
                   end
                 end
               end
@@ -540,7 +540,7 @@ module Punchblock
                   it "should return an error and not execute any actions" do
                     subject.execute
                     error = ProtocolError.new.setup 'option error', 'A repeat_times value is unsupported on Asterisk.'
-                    command.response(0.1).should be == error
+                    original_command.response(0.1).should be == error
                   end
                 end
               end
@@ -559,7 +559,7 @@ module Punchblock
                   it "should return an error and not execute any actions" do
                     subject.execute
                     error = ProtocolError.new.setup 'option error', 'A max_time value is unsupported on Asterisk.'
-                    command.response(0.1).should be == error
+                    original_command.response(0.1).should be == error
                   end
                 end
               end
@@ -578,7 +578,7 @@ module Punchblock
                   it "should return an error and not execute any actions" do
                     subject.execute
                     error = ProtocolError.new.setup 'option error', 'A voice value is unsupported on Asterisk.'
-                    command.response(0.1).should be == error
+                    original_command.response(0.1).should be == error
                   end
                 end
               end
@@ -613,12 +613,64 @@ module Punchblock
                   it "should return an error and not execute any actions" do
                     subject.execute
                     error = ProtocolError.new.setup 'option error', 'An interrupt-on value of speech is unsupported.'
-                    command.response(0.1).should be == error
+                    original_command.response(0.1).should be == error
                   end
                 end
               end
             end
           end
+
+          describe "#execute_command" do
+            context "with a command it does not understand" do
+              let(:command) { Punchblock::Component::Output::Pause.new }
+
+              before { command.request! }
+              it "returns a ProtocolError response" do
+                subject.execute_command command
+                command.response(0.1).should be_a ProtocolError
+              end
+            end
+
+            context "with a Stop command" do
+              let(:command) { Punchblock::Component::Stop.new }
+              let(:reason) { original_command.complete_event(5).reason }
+              let(:channel) { "SIP/1234-00000000" }
+              let :ami_event do
+                RubyAMI::Event.new('AsyncAGI').tap do |e|
+                  e['SubEvent'] = "Start"
+                  e['Channel']  = channel
+                  e['Env']      = "agi_request%3A%20async%0Aagi_channel%3A%20SIP%2F1234-00000000%0Aagi_language%3A%20en%0Aagi_type%3A%20SIP%0Aagi_uniqueid%3A%201320835995.0%0Aagi_version%3A%201.8.4.1%0Aagi_callerid%3A%205678%0Aagi_calleridname%3A%20Jane%20Smith%0Aagi_callingpres%3A%200%0Aagi_callingani2%3A%200%0Aagi_callington%3A%200%0Aagi_callingtns%3A%200%0Aagi_dnid%3A%201000%0Aagi_rdnis%3A%20unknown%0Aagi_context%3A%20default%0Aagi_extension%3A%201000%0Aagi_priority%3A%201%0Aagi_enhanced%3A%200.0%0Aagi_accountcode%3A%20%0Aagi_threadid%3A%204366221312%0A%0A"
+                end
+              end
+
+              before do
+                command.request!
+                original_command.request!
+                original_command.execute!
+              end
+
+              it "sets the command response to true" do
+                mock_call.expects(:redirect_back!)
+                subject.execute_command command
+                command.response(0.1).should be == true
+              end
+
+              it "sends the correct complete event" do
+                mock_call.expects(:redirect_back!)
+                subject.execute_command command
+                original_command.should_not be_complete
+                mock_call.process_ami_event! ami_event
+                reason.should be_a Punchblock::Event::Complete::Stop
+                original_command.should be_complete
+              end
+
+              it "redirects the call by unjoining it" do
+                mock_call.expects(:redirect_back!).with(nil)
+                subject.execute_command command
+              end
+            end
+          end
+
         end
       end
     end